requirements.txt
Flask==3.1.0
Flask_SocketIO==5.5.1
gTTS==2.5.4
numpy==2.2.3
pygame==2.6.1
python-dotenv==1.0.1
PyYAML==6.0.2
rcon==2.4.9
replicate==1.0.4
sty==1.0.6
pulsectl==24.12.0